updated functionality in the ratings and reviews section
