+++ client/components/ratings/characteristics.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const Characteristics = function(props) {
+  return (
+    <div className="quality">
+
+    </div>
+  )
+}
+
+export default Characteristics;
