+++ public/styles.css
@@ -178,7 +178,26 @@ body {
   width: 30%;
   display: inline-block;
 }
+.ratings-title {
+  width: 100%;
+  text-transform: uppercase;
+  font-size: 16px;
+  display: inline-block;
+}
 
+.overall-rating {
+  float: left;
+  height: 45px;
+  font-weight: bold;
+}
+.stars-rating {
+  display: inline-block;
+  padding: 5px 0 20px 10px;
+  height: 20px;
+}
+.percentage-reviews{
+  width: 100%;
+}
 .reviews {
   width: 67%;
   display: inline;
@@ -191,25 +210,32 @@ body {
 }
 .starBox p {
   color: #000;
-  text-transform: uppercase;
-  margin: 0 0 10px;
+  margin: 0;
   padding: 0;
   letter-spacing: 1px;
 }
 .starBox p:nth-child(2) {
   float: right;
   position: relative;
-  top: -30px;
+  top: -20px;
 }
-.star {
-  background: #262626;
+.starbar {
+  background: #aaa;
   padding: 4px;
   box-sizing: border-box;
-  border: 1px solid #0fffb7;
+  border: 1px solid #074b36;
   border-radius: 2px;
 }
-.star-level {
-  background: #0fffb7;
+.starbar-level {
+  background: #074b36;
   width: 100%;
   height:10px;
+}
+.star-shape {
+  display: inline-block;
+  height: 20px;
+  width: 20px;
+  background-color: #999;
+  clip-path: polygon(50% 0%, 63% 38%, 100% 38%, 69% 59%, 82% 100%, 50% 75%, 18% 100%, 31% 59%, 0% 38%, 37% 38%);
+  -webkit-clip-path: polygon(50% 0%, 63% 38%, 100% 38%, 69% 59%, 82% 100%, 50% 75%, 18% 100%, 31% 59%, 0% 38%, 37% 38%);
 }
